Inky Frame: New button test, old is now button demo.
